requirements.txt
Brotli==1.1.0
certifi==2023.7.22
mutagen==1.47.0
pycryptodomex==3.18.0
python-dotenv==1.0.0
websockets==11.0.3
yt-dlp==2023.7.6
